requirements.txt
adversarial-robustness-toolbox==1.18.1
faiss-cpu
openai
flask==3.0.3
grpcio==1.66.1
huggingface-hub==0.26.0
git+https://github.com/nottombrown/imagenet_stubs
jupyter==1.1.1
keras==2.13.1
langchain
lime==0.2.0.1
matplotlib==3.7.5
modelscan==0.7.6
numpy==1.24.0
opencv-python==4.5.1.48
pandas==2.0.3
pillow==10.4.0
scikit-image==0.21.0
scikit-learn==1.3.2
tensorflow==2.13.1
tqdm==4.66.5
transformers==4.45.2